working version with module and main
finally I've found out how to make the module work, i cannot move the
read_sensor_data function to the module, because it initailizes the
variables, but the rest can be moved over. this version could already
be submitted. for the next version i will try to use the proximity
sensor